refactor: replace desktop file validation with configuration checks for protocol registration
