+++ test/driver.test.js
@@ -10,11 +10,9 @@ const { homedir } = require('../src/config/constants');
 const fs = require('fs');
 
 const ProtocolRegistry = require('../src');
-const {
-    checkRegistration,
-    checkDesktopFile
-} = require('./utils/integration-test');
+const { checkRegistration } = require('./utils/integration-test');
 const constants = require('./config/constants');
+const { checkRegistrationConfig } = require('./utils/configuration-test');
 
 const protocol = 'regimen';
 
@@ -106,7 +104,7 @@ test.each([
     async (args) => {
         await ProtocolRegistry.register(protocol, getCommand(), args.options);
         await checkRegistration(protocol, args.options || {});
-        await checkDesktopFile(protocol);
+        await checkRegistrationConfig(protocol);
     },
     constants.jestTimeOut
 );
