+++ test/utils/fileConfig.helpers.js
@@ -1,0 +1,32 @@
+const { join } = require('path');
+const constants = require('../../src/config/constants');
+const fs = require('fs');
+
+const getMimeAppsFilePath = () => {
+    const mimeAppPaths = [
+        join(constants.osHomeDir, '.config', 'mimeapps.list')
+    ];
+
+    return mimeAppPaths;
+};
+
+const getDesktopFileData = (defaultApp) => {
+    const desktopFilePath = join(
+        constants.osHomeDir,
+        '.local/share/applications',
+        defaultApp
+    );
+
+    const fileData = fs.readFileSync(desktopFilePath, 'utf-8');
+    const actualAppName = defaultApp.split('.')[0].replaceAll('_', ' ');
+
+    return {
+        fileData,
+        actualAppName
+    };
+};
+
+module.exports = {
+    getMimeAppsFilePath,
+    getDesktopFileData
+};
